+++ lib/src/model/model.dart
@@ -1,4 +1,4 @@
 export 'src/cmd_plus_mode.dart';
 export 'src/cmd_plus_result.dart';
-export 'src/directory_copy_filter.dart';
-export 'src/file_copy_filter.dart';
+export 'src/directory_filter.dart';
+export 'src/file_filter.dart';